reorganized the project
