--- src/physics/physics.hpp
@@ -1,5 +1,0 @@
-#include "../entities/entities.hpp"
-
-void gravity(Player *player);
-void jumpPhysics(Player *player, MotionState *jumpState);
-void dropdownPhysics(Player *player, MotionState *jumpState);
